pyproject.toml
[project]
name = "discord-qna-archiver"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "django>=6.0",
    "django-import-export>=4.3.14",
    "google-generativeai>=0.8.6",
    "pillow>=12.0.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
